README.md: Change the link of ImageNet2012 data manual
Currently, tensorflow/models repository's slim bazel BUILD file is
broken and not working (tensorflow/models#9539). Using the old BUILD
file is not working too due to change of the URL. So I forked the
tensorflow/models project and updated the bazel BUILD file just for this
project, and changing the Link might be helpful for the other
developers.

Reference:
  https://github.com/tensorflow/models/issues/9539
  https://github.com/simnalamburt/models/commit/b95a173087cbc30394dc54eec1a2cf0033c9f173